requirements.txt
Flask==3.0.0
gunicorn==21.2.0
tensorflow==2.10.0
numpy==1.23.5
Pillow==10.0.0
reportlab==4.0.4
